only run qemu image when host machine is not ARM64

+++ ci-tests.sh
@@ -12,7 +12,9 @@ IMAGE_TAG="latest"
 
 if [ "$ARM" = "1" ] || [ "$( uname -m )" = "aarch64" ]; then
   ARCH="--platform=linux/arm64"
-  docker run --rm --privileged multiarch/qemu-user-static --reset -p yes # enable qemu support for docker
+  if [ "$( uname -m )" != "aarch64" ]; then
+    docker run --rm --privileged multiarch/qemu-user-static --reset -p yes # enable qemu support for docker
+  fi
 else
   ARCH="--platform=linux/amd64"
 fi
